Shorten message history in online mode
